LayoutTests/editing/pasteboard/paste-text-015-expected.txt

This test copies, deletes, and paste back <div id="test" class="editing"><p>foo</p><p>bar</p></div>
It then inserts the paragraph and pastes the contents again.
| "\n"
| <div>
|   class="editing"
|   id="test"
|   <p>
|     "foo"
|   <p>
|     "bar"
|   <p>
|     "foo"
|   <p>
|     "bar<#selection-caret>"
| "\n"
